solucion de problemas

+++ proyectoFinal/src/app/pages/admin-users/admin-users.component.ts
@@ -8,11 +8,10 @@ import Swal from 'sweetalert2';
 import { Router } from '@angular/router';
 import { AuthService } from 'src/app/shared/auth.service';
 
-
 @Component({
   selector: 'app-admin-users',
   templateUrl: './admin-users.component.html',
-  styleUrls: ['./admin-users.component.css']
+  styleUrls: ['./admin-users.component.css'],
 })
 export class AdminUsersComponent implements OnInit {
   @ViewChild('deleteAdmin') deleteAdmin;
@@ -25,14 +24,46 @@ export class AdminUsersComponent implements OnInit {
   public userall: User[];
   public myIndex: number;
 
-  constructor(public userService: UserService, private formBuilder: FormBuilder, public EncrDecr: EncrDecrServiceService, private router: Router, private serviceTitle: Title, private auth: AuthService) {
+  constructor(
+    public userService: UserService,
+    private formBuilder: FormBuilder,
+    public EncrDecr: EncrDecrServiceService,
+    private router: Router,
+    private serviceTitle: Title,
+    private auth: AuthService
+  ) {
     this.title = 'ADM - USUARIOS';
-    this.nationalitiesEdit = ['Panama', 'Colombia', 'Costa Rica', 'Honduras', 'Brazil', 'Argentina', 'Bolivia', 'Cuba', 'El Salvador', 'Ecuador', 'Guatemala', 'Jamaica', 'Mexico', 'Nicaragua', 'Paraguay', 'Peru', 'Puerto Rico', 'Espana', 'Estados Unidos', 'Uruguay', 'Venezuela', 'Portugal', 'China', 'Republica Dominicana'];
+    this.nationalitiesEdit = [
+      'Panama',
+      'Colombia',
+      'Costa Rica',
+      'Honduras',
+      'Brazil',
+      'Argentina',
+      'Bolivia',
+      'Cuba',
+      'El Salvador',
+      'Ecuador',
+      'Guatemala',
+      'Jamaica',
+      'Mexico',
+      'Nicaragua',
+      'Paraguay',
+      'Peru',
+      'Puerto Rico',
+      'Espana',
+      'Estados Unidos',
+      'Uruguay',
+      'Venezuela',
+      'Portugal',
+      'China',
+      'Republica Dominicana',
+    ];
     this.typeUser = 'user';
     this.userlogin = false;
     this.adminlogin = false;
     this.g2pUserPerfil = this.userService.usuarios;
-    this.userall = this.userService.collection
+    this.userall = this.userService.collection;
     this.myIndex = 0;
   }
 
@@ -44,45 +75,55 @@ export class AdminUsersComponent implements OnInit {
   adminUser() {
     this.userService.getUserAll().subscribe((data: User[]) => {
       this.userService.collection = data;
-      this.userall = this.userService.collection
+      this.userall = this.userService.collection;
       localStorage.setItem('adminuser', JSON.stringify(this.userall));
-    })
+    });
   }
 
   deleteUser(id: any, id2: any) {
     this.userService.deleteUser(parseInt(id)).subscribe((data) => {
-      this.deleteAdmin.nativeElement.click();
-      this.userall.splice(id2, 1);
-      localStorage.setItem('adminuser', JSON.stringify(this.userall))
-      Swal.fire({
-        position: 'center',
-        icon: 'success',
-        title: 'Perfil eliminado correctamente!',
-        showConfirmButton: false,
-        timer: 2500
-      });
-    })
+      if (data === true) {
+        Swal.fire({
+          position: 'center',
+          icon: 'error',
+          title: 'Ha ocurrido un error, por favor solucionalo',
+          showConfirmButton: false,
+          timer: 2500,
+        });
+      } else {
+        this.deleteAdmin.nativeElement.click();
+        this.userall.splice(id2, 1);
+        localStorage.setItem('adminuser', JSON.stringify(this.userall));
+        Swal.fire({
+          position: 'center',
+          icon: 'success',
+          title: 'Perfil eliminado correctamente!',
+          showConfirmButton: false,
+          timer: 2500,
+        });
+      }
+    });
     // window.location.reload()
   }
 
   public displayInfo(i: number) {
-    this.myIndex = i
+    this.myIndex = i;
   }
 
   noAdmin() {
-    if (this.isLoggedIn() && this.g2pUserPerfil.admin === "admin") {
-      console.log("hola");
-      return true
+    if (this.isLoggedIn() && this.g2pUserPerfil.admin === 'admin') {
+      console.log('hola');
+      return true;
     } else {
       Swal.fire({
         position: 'center',
         icon: 'error',
         title: 'No tienes permiso para entrar aqui',
         showConfirmButton: false,
-        timer: 2000
+        timer: 2000,
       });
       this.router.navigateByUrl('/');
-      return false
+      return false;
     }
   }
